requirements.txt
asgiref==3.9.1
diff-match-patch==20241021
Django==5.2.5
django-import-export==4.3.9
django-jazzmin==3.0.1
django-modeltranslation==0.19.16
djangorestframework==3.16.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
inflection==0.5.1
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
python-environ==0.4.54
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.14.1
uritemplate==4.2.0
